+++ be/src/udf/udf.h
@@ -47,7 +47,6 @@ namespace doris_udf {
 // object containing a boolean to store if the value is nullptr and the value itself. The
 // value is unspecified if the nullptr boolean is set.
 struct AnyVal;
-struct BooleanVal;
 struct StringVal;
 struct DateTimeVal;
 struct DecimalV2Val;
@@ -265,133 +264,6 @@ private:
 };
 
 //----------------------------------------------------------------------------
-//------------------------------- UDFs ---------------------------------------
-//----------------------------------------------------------------------------
-// The UDF function must implement this function prototype. This is not
-// a typedef as the actual UDF's signature varies from UDF to UDF.
-//    typedef <*Val> Evaluate(FunctionContext* context, <const Val& arg>);
-//
-// The UDF must return one of the *Val structs. The UDF must accept a pointer
-// to a FunctionContext object and then a const reference for each of the input arguments.
-// nullptr input arguments will have nullptr passed in.
-// Examples of valid Udf signatures are:
-//  1) DoubleVal Example1(FunctionContext* context);
-//  2) IntVal Example2(FunctionContext* context, const IntVal& a1, const DoubleVal& a2);
-//
-// UDFs can be variadic. The variable arguments must all come at the end and must be
-// the same type. A example signature is:
-//  StringVal Concat(FunctionContext* context, const StringVal& separator,
-//    int num_var_args, const StringVal* args);
-// In this case args[0] is the first variable argument and args[num_var_args - 1] is
-// the last.
-//
-// The UDF should not maintain any state across calls since there is no guarantee
-// on how the execution is multithreaded or distributed. Conceptually, the UDF
-// should only read the input arguments and return the result, using only the
-// FunctionContext as an external object.
-//
-// Memory Management: the UDF can assume that memory from input arguments will have
-// the same lifetime as results for the UDF. In other words, the UDF can return
-// memory from input arguments without making copies. For example, a function like
-// substring will not need to allocate and copy the smaller string. For cases where
-// the UDF needs a buffer, it should use the StringRef(FunctionContext, len) c'tor.
-// TODO: things above is not right. StringRef shouldn't use here.
-//
-// The UDF can optionally specify a Prepare function. The prepare function is called
-// once before any calls to the Udf to evaluate values. This is the appropriate time for
-// the Udf to validate versions and things like that.
-// If there is an error, this function should call FunctionContext::set_error()/
-// FunctionContext::add_warning().
-typedef void (*UdfPrepareFn)(FunctionContext* context);
-
-/// --- Prepare / Close Functions ---
-/// ---------------------------------
-/// The UDF can optionally include a prepare function, specified in the "CREATE FUNCTION"
-/// statement using "prepare_fn=<prepare function symbol>". The prepare function is called
-/// before any calls to the UDF to evaluate values. This is the appropriate time for the
-/// UDF to initialize any shared data structures, validate versions, etc. If there is an
-/// error, this function should call FunctionContext::SetError()/
-/// FunctionContext::AddWarning().
-//
-/// The prepare function is called multiple times with different FunctionStateScopes. It
-/// will be called once per fragment with 'scope' set to FRAGMENT_LOCAL, and once per
-/// execution thread with 'scope' set to THREAD_LOCAL.
-typedef void (*UdfPrepare)(FunctionContext* context, FunctionContext::FunctionStateScope scope);
-
-/// The UDF can also optionally include a close function, specified in the "CREATE
-/// FUNCTION" statement using "close_fn=<close function symbol>". The close function is
-/// called after all calls to the UDF have completed. This is the appropriate time for the
-/// UDF to deallocate any shared data structures that are not needed to maintain the
-/// results. If there is an error, this function should call FunctionContext::SetError()/
-/// FunctionContext::AddWarning().
-//
-/// The close function is called multiple times with different FunctionStateScopes. It
-/// will be called once per fragment with 'scope' set to FRAGMENT_LOCAL, and once per
-/// execution thread with 'scope' set to THREAD_LOCAL.
-typedef void (*UdfClose)(FunctionContext* context, FunctionContext::FunctionStateScope scope);
-
-//----------------------------------------------------------------------------
-//------------------------------- UDAs ---------------------------------------
-//----------------------------------------------------------------------------
-// The UDA execution is broken up into a few steps. The general calling pattern
-// is one of these:
-//  1) Init(), Evaluate() (repeatedly), Serialize()
-//  2) Init(), Merge() (repeatedly), Serialize()
-//  3) Init(), Finalize()
-// The UDA is registered with three types: the result type, the input type and
-// the intermediate type.
-//
-// If the UDA needs a fixed byte width intermediate buffer, the type should be
-// TYPE_FIXED_BUFFER and Doris will allocate the buffer. If the UDA needs an unknown
-// sized buffer, it should use TYPE_STRING and allocate it from the FunctionContext
-// manually.
-// For UDAs that need a complex data structure as the intermediate state, the
-// intermediate type should be string and the UDA can cast the ptr to the structure
-// it is using.
-//
-// Memory Management: For allocations that are not returned to Doris, the UDA
-// should use the FunctionContext::Allocate()/Free() methods. For StringVal allocations
-// returned to Doris (e.g. UdaSerialize()), the UDA should allocate the result
-// via StringVal(FunctionContext*, int) ctor and Doris will automatically handle
-// freeing it.
-//
-// For clarity in documenting the UDA interface, the various types will be typedefed
-// here. The actual execution resolves all the types at runtime and none of these types
-// should actually be used.
-typedef AnyVal InputType;
-typedef AnyVal InputType2;
-typedef AnyVal ResultType;
-typedef AnyVal IntermediateType;
-
-// UdaInit is called once for each aggregate group before calls to any of the
-// other functions below.
-typedef void (*UdaInit)(FunctionContext* context, IntermediateType* result);
-
-// This is called for each input value. The UDA should update result based on the
-// input value. The update function can take any number of input arguments. Here
-// are some examples:
-typedef void (*UdaUpdate)(FunctionContext* context, const InputType& input,
-                          IntermediateType* result);
-typedef void (*UdaUpdate2)(FunctionContext* context, const InputType& input,
-                           const InputType2& input2, IntermediateType* result);
-
-// Merge an intermediate result 'src' into 'dst'.
-typedef void (*UdaMerge)(FunctionContext* context, const IntermediateType& src,
-                         IntermediateType* dst);
-
-// Serialize the intermediate type. The serialized data is then sent across the
-// wire. This is not called unless the intermediate type is String.
-// No additional functions will be called with this FunctionContext object and the
-// UDA should do final clean (e.g. Free()) here.
-typedef const IntermediateType (*UdaSerialize)(FunctionContext* context,
-                                               const IntermediateType& type);
-
-// Called once at the end to return the final value for this UDA.
-// No additional functions will be called with this FunctionContext object and the
-// UDA should do final clean (e.g. Free()) here.
-typedef ResultType (*UdaFinalize)(FunctionContext* context, const IntermediateType& v);
-
-//----------------------------------------------------------------------------
 //-------------Implementation of the *Val structs ----------------------------
 //----------------------------------------------------------------------------
 struct AnyVal {
@@ -399,33 +271,6 @@ struct AnyVal {
     AnyVal() : is_null(false) {}
     AnyVal(bool is_null) : is_null(is_null) {}
 };
-
-struct BooleanVal : public AnyVal {
-    bool val;
-
-    BooleanVal() : val(false) {}
-    BooleanVal(bool val) : val(val) {}
-
-    static BooleanVal null() {
-        BooleanVal result;
-        result.is_null = true;
-        return result;
-    }
-
-    bool operator==(const BooleanVal& other) const {
-        if (is_null && other.is_null) {
-            return true;
-        }
-
-        if (is_null || other.is_null) {
-            return false;
-        }
-
-        return val == other.val;
-    }
-    bool operator!=(const BooleanVal& other) const { return !(*this == other); }
-};
-
 struct BigIntVal : public AnyVal {
     int64_t val;
 
@@ -643,7 +488,6 @@ struct DecimalV2Val : public AnyVal {
     bool operator!=(const DecimalV2Val& other) const { return !(*this == other); }
 };
 
-using doris_udf::BooleanVal;
 using doris_udf::BigIntVal;
 using doris_udf::DoubleVal;
 using doris_udf::StringVal;
